#1256 Move GOConstLoopBounds into its own file in domain/gocean/transformations
